Set array size only after allocating the new memory portion (#58848)
The `.size` of an array might be getting set too early; before the
memory is actually allocated, resulting in larger race windows for
inconsistencies (in particular, during finalizers on this thread) to
manifest. This helps keep those instructions more nearly consecutive to
each other so there is a smaller window for races to be visible, in
typical non-adversarial situations.